requirements.txt
Django==4.0.6
django-admin-sortable2==2.1
django-tinymce==3.4.0
idna==3.3
Pillow==9.2.0
python-dotenv==0.20.0
requests==2.28.1

